Recompiled static website to implement redesign.
