+++ docs/404.html
@@ -1,30 +1,30 @@
 <!doctype html>
-<html lang="en" data-theme="system">
+<html lang="en">
 	<head>
 		<meta charset="utf-8" />
 		<link rel="icon" href="/favicon.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1, viewport-fit=cover" />
 		<meta name="disabled-adaptations" content="watch" />
 		
-		<link rel="modulepreload" href="/_app/immutable/entry/start.Cnf-AIQw.js">
-		<link rel="modulepreload" href="/_app/immutable/chunks/entry.D9By3_zt.js">
+		<link rel="modulepreload" href="/_app/immutable/entry/start.Dyr0b8xi.js">
+		<link rel="modulepreload" href="/_app/immutable/chunks/entry.BIxGs3Gi.js">
 		<link rel="modulepreload" href="/_app/immutable/chunks/scheduler.LR2eQXLY.js">
-		<link rel="modulepreload" href="/_app/immutable/entry/app.DwiEUvbT.js">
+		<link rel="modulepreload" href="/_app/immutable/entry/app.BOUzON3X.js">
 		<link rel="modulepreload" href="/_app/immutable/chunks/index.weWyDj9-.js">
 	</head>
 	<body data-sveltekit-preload-data="hover" id="top">
 		<div style="display: contents">
 			<script>
 				{
-					__sveltekit_13bmcl1 = {
+					__sveltekit_rr8p75 = {
 						base: ""
 					};
 
 					const element = document.currentScript.parentElement;
 
 					Promise.all([
-						import("/_app/immutable/entry/start.Cnf-AIQw.js"),
-						import("/_app/immutable/entry/app.DwiEUvbT.js")
+						import("/_app/immutable/entry/start.Dyr0b8xi.js"),
+						import("/_app/immutable/entry/app.BOUzON3X.js")
 					]).then(([kit, app]) => {
 						kit.start(app, element);
 					});
