--- docs/_app/immutable/assets/1.AcUXp8PE.css
@@ -1,1 +1,0 @@
-.container.svelte-1361tq6{position:relative;display:block;font-weight:900;text-align:center;margin:auto}.status.svelte-1361tq6{position:absolute;top:50%;left:50%;z-index:var(--z-underlay);transform:translate(-50%,-48%);display:block;font-family:var(--font-display);font-size:max(var(--font-4xl),min(60cqi,60cqb));-webkit-user-select:none;-moz-user-select:none;user-select:none;opacity:.05}.message.svelte-1361tq6{display:block;font-size:var(--font-4xl);margin:0}
