+++ docs/_app/immutable/assets/_error.BmmAZKQf.css
@@ -1,0 +1,1 @@
+.container.svelte-16mor8m{position:relative;display:block;text-align:center;margin:auto}.status.svelte-16mor8m{position:absolute;top:50%;left:50%;z-index:var(--z-underlay);transform:translate(-50%,-48%);display:block;font-family:var(--font-display);font-size:max(var(--font-4xl),min(60cqi,60cqb));font-weight:900;-webkit-user-select:none;-moz-user-select:none;user-select:none;opacity:.05}.message.svelte-16mor8m{display:block;font-size:var(--font-4xl);margin:0}
