+++ docs/_app/immutable/assets/_page.DB3wxkKI.css
@@ -1,0 +1,1 @@
+h1.svelte-ai4q0w.svelte-ai4q0w{display:block;font-size:var(--font-3xl);margin:0}.grid.svelte-ai4q0w.svelte-ai4q0w{--gap:var(--space-xs-s);--min:var(--width-xs);--max:calc(100% / 3);display:grid;grid-template-columns:repeat(auto-fill,minmax(clamp(calc(100% / 3) - var(--space-xs-s),var(--width-xs),100%),1fr));grid-template-columns:repeat(auto-fill,minmax(clamp(var(--max) - var(--gap),var(--min),100%),1fr));align-items:stretch;grid-gap:var(--space-xs-s);grid-gap:var(--gap);gap:var(--space-xs-s);gap:var(--gap);margin:var(--space-m-l) 0}.project.svelte-ai4q0w.svelte-ai4q0w{display:flex;flex-flow:column nowrap;justify-content:flex-start;container-type:inline-size;border:2px dashed var(--bg-color-3);border-radius:var(--space-3xs-2xs);background-color:var(--bg-color-1);padding:var(--space-s) 0;gap:var(--space-xs-s) 0}.project.svelte-ai4q0w .title.svelte-ai4q0w,.project.svelte-ai4q0w .portal.svelte-ai4q0w,.project.svelte-ai4q0w .description.svelte-ai4q0w,.project.svelte-ai4q0w .learn-more.svelte-ai4q0w{display:block;font-size:var(--font-m);margin:0 var(--space-s)}.project.svelte-ai4q0w .title.svelte-ai4q0w{font-size:var(--font-2xl)}.project.svelte-ai4q0w .portal.svelte-ai4q0w{position:relative;z-index:var(--z-overlay);aspect-ratio:3 / 2;width:100%;margin:0;transition:transform .1s ease-out;border:1px dashed var(--bg-color-3);background-color:inherit;color:var(--color);overflow:hidden}.project.svelte-ai4q0w .portal.svelte-ai4q0w:hover,.project.svelte-ai4q0w .portal.svelte-ai4q0w:focus{border-radius:inherit;transform:scale(1.03);color:var(--color-accent)}.project.svelte-ai4q0w .portal.svelte-ai4q0w,.project.svelte-ai4q0w .portal.svelte-ai4q0w:active{transform:none}.project.svelte-ai4q0w .portal img.svelte-ai4q0w,.project.svelte-ai4q0w .portal svg.svelte-ai4q0w{display:block;width:100%;height:100%;-o-object-fit:cover;object-fit:cover;-o-object-position:top center;object-position:top center}.project.svelte-ai4q0w .learn-more.svelte-ai4q0w{text-align:right;border-top:1px solid var(--bg-color-2);padding-top:inherit;margin-top:auto}
