--- docs/_app/immutable/assets/_page.DLfm78fW.css
@@ -1,1 +1,0 @@
-h1.svelte-1ymw1bw.svelte-1ymw1bw{display:block;font-size:var(--font-3xl);margin:0}.grid.svelte-1ymw1bw.svelte-1ymw1bw{--gap:var(--space-m-l);--min:var(--width-xs);--max:calc(100% / 3);display:grid;grid-template-columns:repeat(auto-fill,minmax(clamp(calc(100% / 3) - var(--space-m-l),var(--width-xs),100%),1fr));grid-template-columns:repeat(auto-fill,minmax(clamp(var(--max) - var(--gap),var(--min),100%),1fr));align-items:stretch;grid-gap:var(--space-m-l);grid-gap:var(--gap);gap:var(--space-m-l);gap:var(--gap);margin:var(--space-m-l) 0}.project.svelte-1ymw1bw.svelte-1ymw1bw{display:flex;flex-flow:column nowrap;justify-content:flex-start;container-type:inline-size;border-radius:3px;background-color:var(--bg-color-1);box-shadow:0 4px 16px -12px #000;padding:var(--space-s) 0;gap:var(--space-xs-s) 0}.project.svelte-1ymw1bw .title.svelte-1ymw1bw,.project.svelte-1ymw1bw .portal.svelte-1ymw1bw,.project.svelte-1ymw1bw .description.svelte-1ymw1bw,.project.svelte-1ymw1bw .learn-more.svelte-1ymw1bw{display:block;font-size:var(--font-m);margin:0 var(--space-s)}.project.svelte-1ymw1bw .title.svelte-1ymw1bw{font-size:var(--font-2xl)}.project.svelte-1ymw1bw .portal.svelte-1ymw1bw{position:relative;z-index:var(--z-overlay);aspect-ratio:3 / 2;width:100%;margin:0;transition:transform .1s ease-out;background-color:var(--bg-color-2);background-image:radial-gradient(transparent,var(--bg-color-1));color:var(--color);overflow:hidden}.project.svelte-1ymw1bw .portal.svelte-1ymw1bw:hover,.project.svelte-1ymw1bw .portal.svelte-1ymw1bw:focus{border-radius:inherit;transform:scale(1.03);color:var(--color-accent)}.project.svelte-1ymw1bw .portal.svelte-1ymw1bw,.project.svelte-1ymw1bw .portal.svelte-1ymw1bw:active{transform:none}.project.svelte-1ymw1bw .portal img.svelte-1ymw1bw,.project.svelte-1ymw1bw .portal svg.svelte-1ymw1bw{display:block;width:100%;height:100%;-o-object-fit:cover;object-fit:cover;-o-object-position:top center;object-position:top center}.project.svelte-1ymw1bw .learn-more.svelte-1ymw1bw{text-align:right;border-top:1px solid var(--bg-color-2);padding-top:inherit;margin-top:auto}
