+++ docs/_app/immutable/chunks/Meta.DZWJ86d6.js
@@ -1,0 +1,1 @@
+import{s as _,n as g,c as M}from"./scheduler.LR2eQXLY.js";import{S as E,i as A,e as d,h as T,c as l,d as s,a as e,b as h}from"./index.weWyDj9-.js";import{p as v}from"./stores.CgJopnDC.js";function b(c){let m,o,p,a,r,n,y,u,f;return document.title=m=c[0]+" | Hey, Jes!",{c(){o=d("meta"),p=d("meta"),a=d("meta"),r=d("meta"),n=d("meta"),y=d("meta"),u=d("meta"),this.h()},l(i){const t=T("svelte-159ivmp",document.head);o=l(t,"META",{name:!0,content:!0}),p=l(t,"META",{property:!0,content:!0}),a=l(t,"META",{property:!0,content:!0}),r=l(t,"META",{property:!0,content:!0}),n=l(t,"META",{property:!0,content:!0}),y=l(t,"META",{property:!0,content:!0}),u=l(t,"META",{property:!0,content:!0}),t.forEach(s),this.h()},h(){e(o,"name","description"),e(o,"content",c[1]),e(p,"property","og:site_name"),e(p,"content","Hey, Jes!"),e(a,"property","og:title"),e(a,"content",c[0]),e(r,"property","og:description"),e(r,"content",c[1]),e(n,"property","og:type"),e(n,"content","website"),e(y,"property","og:image"),e(y,"content","/opengraph.png"),e(u,"property","og:url"),e(u,"content",f="https://heyjes.dev"+c[2].url.pathname)},m(i,t){h(document.head,o),h(document.head,p),h(document.head,a),h(document.head,r),h(document.head,n),h(document.head,y),h(document.head,u)},p(i,[t]){t&1&&m!==(m=i[0]+" | Hey, Jes!")&&(document.title=m),t&2&&e(o,"content",i[1]),t&1&&e(a,"content",i[0]),t&2&&e(r,"content",i[1]),t&4&&f!==(f="https://heyjes.dev"+i[2].url.pathname)&&e(u,"content",f)},i:g,o:g,d(i){s(o),s(p),s(a),s(r),s(n),s(y),s(u)}}}function H(c,m,o){let p;M(c,v,n=>o(2,p=n));let{title:a}=m,{description:r}=m;return c.$$set=n=>{"title"in n&&o(0,a=n.title),"description"in n&&o(1,r=n.description)},[a,r,p]}class q extends E{constructor(m){super(),A(this,m,H,b,_,{title:0,description:1})}}export{q as M};
