+++ docs/_app/immutable/chunks/stores.CgJopnDC.js
@@ -1,0 +1,1 @@
+import{s as e}from"./entry.BIxGs3Gi.js";const r=()=>{const s=e;return{page:{subscribe:s.page.subscribe},navigating:{subscribe:s.navigating.subscribe},updated:s.updated}},b={subscribe(s){return r().page.subscribe(s)}};export{b as p};
