--- docs/_app/immutable/chunks/stores.D6sMEoti.js
@@ -1,1 +1,0 @@
-import{s as e}from"./entry.D9By3_zt.js";const r=()=>{const s=e;return{page:{subscribe:s.page.subscribe},navigating:{subscribe:s.navigating.subscribe},updated:s.updated}},b={subscribe(s){return r().page.subscribe(s)}};export{b as p};
