--- docs/_app/immutable/nodes/0.GHqBXa_A.js
@@ -1,1 +1,0 @@
-import{s as P,n as D,c as U,o as K,b as Q,u as W,v as X,w as Z,x as ee}from"../chunks/scheduler.LR2eQXLY.js";import{S as F,i as J,e as m,s as w,c as p,k as S,m as T,j as C,d as j,a as o,D as L,o as A,b as i,t as I,l as N,f as B,g as R,n as Y,q as O,r as V,u as G}from"../chunks/index.weWyDj9-.js";import{p as te}from"../chunks/stores.D6sMEoti.js";/* empty css                                    */const se=!0,me=Object.freeze(Object.defineProperty({__proto__:null,prerender:se},Symbol.toStringTag,{value:"Module"}));function ae(f){let s,a,l,n='<a href="/" class="svelte-1qj3bqz">Hey, Jes!</a>',u,r,d,c="Home",e,t,k="About",x,h,H="Projects",y,v,z="Github",$,b,E="Linkedin";return{c(){s=m("header"),a=m("div"),l=m("h2"),l.innerHTML=n,u=w(),r=m("nav"),d=m("a"),d.textContent=c,e=w(),t=m("a"),t.textContent=k,x=w(),h=m("a"),h.textContent=H,y=w(),v=m("a"),v.textContent=z,$=w(),b=m("a"),b.textContent=E,this.h()},l(_){s=p(_,"HEADER",{class:!0});var g=S(s);a=p(g,"DIV",{class:!0});var M=S(a);l=p(M,"H2",{class:!0,"data-svelte-h":!0}),T(l)!=="svelte-1t7c000"&&(l.innerHTML=n),u=C(M),r=p(M,"NAV",{id:!0,class:!0});var q=S(r);d=p(q,"A",{href:!0,class:!0,"data-svelte-h":!0}),T(d)!=="svelte-peh6sg"&&(d.textContent=c),e=C(q),t=p(q,"A",{href:!0,class:!0,"data-svelte-h":!0}),T(t)!=="svelte-85qk6k"&&(t.textContent=k),x=C(q),h=p(q,"A",{href:!0,class:!0,"data-svelte-h":!0}),T(h)!=="svelte-j4svvd"&&(h.textContent=H),y=C(q),v=p(q,"A",{href:!0,target:!0,class:!0,"data-svelte-h":!0}),T(v)!=="svelte-h1iwbv"&&(v.textContent=z),$=C(q),b=p(q,"A",{href:!0,target:!0,class:!0,"data-svelte-h":!0}),T(b)!=="svelte-1at6uta"&&(b.textContent=E),q.forEach(j),M.forEach(j),g.forEach(j),this.h()},h(){o(l,"class","brand svelte-1qj3bqz"),o(d,"href","/"),o(d,"class","svelte-1qj3bqz"),L(d,"active",f[0]==="/"),o(t,"href","/about"),o(t,"class","svelte-1qj3bqz"),L(t,"active",f[0]==="/about"),o(h,"href","/projects"),o(h,"class","svelte-1qj3bqz"),L(h,"active",f[0]==="/projects"),o(v,"href","https://github.com/joshuasand0val"),o(v,"target","_blank"),o(v,"class","svelte-1qj3bqz"),o(b,"href","https://linkedin.com/in/joshuasand0val/"),o(b,"target","_blank"),o(b,"class","svelte-1qj3bqz"),o(r,"id","nav"),o(r,"class","svelte-1qj3bqz"),o(a,"class","wrapper svelte-1qj3bqz"),o(s,"class","container svelte-1qj3bqz")},m(_,g){A(_,s,g),i(s,a),i(a,l),i(a,u),i(a,r),i(r,d),i(r,e),i(r,t),i(r,x),i(r,h),i(r,y),i(r,v),i(r,$),i(r,b)},p(_,[g]){g&1&&L(d,"active",_[0]==="/"),g&1&&L(t,"active",_[0]==="/about"),g&1&&L(h,"active",_[0]==="/projects")},i:D,o:D,d(_){_&&j(s)}}}function le(f,s,a){let l,n;return U(f,te,u=>a(1,n=u)),f.$$.update=()=>{f.$$.dirty&2&&a(0,l=n.url.pathname)},[l,n]}class re extends F{constructor(s){super(),J(this,s,le,ae,P,{})}}function ne(f){let s,a,l,n,u,r=new Date().getFullYear()+"",d,c,e,t,k='Site built with <a target="_blank" href="https://svelte.dev">Svelte</a>.',x,h,H='<i class="bi bi-arrow-up-circle-fill"></i>',y,v,z='<i class="system bi bi-circle-half" aria-label="Swap from system theme"></i> <i class="light bi bi-brightness-high-fill" aria-label="Swap from light theme"></i> <i class="dark bi bi-moon-stars-fill" aria-label="Swap from dark theme"></i>';return{c(){s=m("footer"),a=m("div"),l=m("div"),n=m("span"),u=I("© "),d=I(r),c=I(" Joshua Elijah Sandoval."),e=w(),t=m("span"),t.innerHTML=k,x=w(),h=m("a"),h.innerHTML=H,y=w(),v=m("button"),v.innerHTML=z,this.h()},l($){s=p($,"FOOTER",{class:!0});var b=S(s);a=p(b,"DIV",{class:!0});var E=S(a);l=p(E,"DIV",{class:!0});var _=S(l);n=p(_,"SPAN",{class:!0});var g=S(n);u=N(g,"© "),d=N(g,r),c=N(g," Joshua Elijah Sandoval."),g.forEach(j),e=C(_),t=p(_,"SPAN",{class:!0,"data-svelte-h":!0}),T(t)!=="svelte-1sf0pv9"&&(t.innerHTML=k),_.forEach(j),x=C(E),h=p(E,"A",{href:!0,class:!0,"aria-label":!0,"data-svelte-h":!0}),T(h)!=="svelte-1rnzisb"&&(h.innerHTML=H),y=C(E),v=p(E,"BUTTON",{"aria-label":!0,class:!0,"data-svelte-h":!0}),T(v)!=="svelte-p64boq"&&(v.innerHTML=z),E.forEach(j),b.forEach(j),this.h()},h(){o(n,"class","svelte-a0jcg2"),o(t,"class","svelte-a0jcg2"),o(l,"class","info svelte-a0jcg2"),o(h,"href","#top"),o(h,"class","top svelte-a0jcg2"),o(h,"aria-label","To top of page"),L(h,"hidden",!f[0]),o(v,"aria-label","Swap current theme"),o(v,"class","toggle svelte-a0jcg2"),o(a,"class","wrapper svelte-a0jcg2"),o(s,"class","container svelte-a0jcg2")},m($,b){A($,s,b),i(s,a),i(a,l),i(l,n),i(n,u),i(n,d),i(n,c),i(l,e),i(l,t),i(a,x),i(a,h),i(a,y),i(a,v),f[2](v)},p($,[b]){b&1&&L(h,"hidden",!$[0])},i:D,o:D,d($){$&&j(s),f[2](null)}}}function oe(f,s,a){let l=!1,n;K(()=>{const r=document.documentElement,d=()=>{a(0,l=r.scrollHeight>r.offsetHeight&&window.scrollY!==0)};document.addEventListener("scroll",d),window.addEventListener("resize",d);const c=()=>localStorage.getItem("theme")??"system",e=()=>r.setAttribute("data-theme",c());e(),window.addEventListener("storage",e);const t=()=>window.matchMedia("(prefers-color-scheme: light)").matches;n.addEventListener("click",()=>{const k=(()=>{switch(c()){case"system":return t()?"light":"dark";case"light":return t()?"dark":"system";case"dark":return t()?"system":"light";default:return"system"}})();localStorage.setItem("theme",k),e()})});function u(r){Q[r?"unshift":"push"](()=>{n=r,a(1,n)})}return[l,n,u]}class ie extends F{constructor(s){super(),J(this,s,oe,ne,P,{})}}function ce(f){let s,a,l,n,u,r;s=new re({});const d=f[1].default,c=W(d,f,f[0],null);return u=new ie({}),{c(){B(s.$$.fragment),a=w(),l=m("main"),c&&c.c(),n=w(),B(u.$$.fragment),this.h()},l(e){R(s.$$.fragment,e),a=C(e),l=p(e,"MAIN",{class:!0});var t=S(l);c&&c.l(t),t.forEach(j),n=C(e),R(u.$$.fragment,e),this.h()},h(){o(l,"class","svelte-gduh3g")},m(e,t){Y(s,e,t),A(e,a,t),A(e,l,t),c&&c.m(l,null),A(e,n,t),Y(u,e,t),r=!0},p(e,[t]){c&&c.p&&(!r||t&1)&&X(c,d,e,e[0],r?ee(d,e[0],t,null):Z(e[0]),null)},i(e){r||(O(s.$$.fragment,e),O(c,e),O(u.$$.fragment,e),r=!0)},o(e){V(s.$$.fragment,e),V(c,e),V(u.$$.fragment,e),r=!1},d(e){e&&(j(a),j(l),j(n)),G(s,e),c&&c.d(e),G(u,e)}}}function ue(f,s,a){let{$$slots:l={},$$scope:n}=s;return f.$$set=u=>{"$$scope"in u&&a(0,n=u.$$scope)},[n,l]}class pe extends F{constructor(s){super(),J(this,s,ue,ce,P,{})}}export{pe as component,me as universal};
