+++ docs/_app/immutable/nodes/0.yDhr_A5_.js
@@ -1,0 +1,1 @@
+import{s as q,n as N,c as K,o as Q,u as U,v as W,w as X,x as Z}from"../chunks/scheduler.LR2eQXLY.js";import{S as z,i as F,e as p,s as C,c as m,k as H,m as w,j as y,d as $,a as r,D as A,o as k,b as c,t as V,l as I,f as R,g as Y,n as G,q as O,r as P,u as B}from"../chunks/index.weWyDj9-.js";import{p as ee}from"../chunks/stores.CgJopnDC.js";/* empty css                                    */const te=!0,de=Object.freeze(Object.defineProperty({__proto__:null,prerender:te},Symbol.toStringTag,{value:"Module"}));function ae(v){let t,l,a,n='<a href="/" class="svelte-jmayj8">Hey, Jes!</a>',i,o,h,f="Home",e,s,M="About",L,u,T="Projects",j,d,S="Github",x,_,J="Linkedin";return{c(){t=p("header"),l=p("div"),a=p("h2"),a.innerHTML=n,i=C(),o=p("nav"),h=p("a"),h.textContent=f,e=C(),s=p("a"),s.textContent=M,L=C(),u=p("a"),u.textContent=T,j=C(),d=p("a"),d.textContent=S,x=C(),_=p("a"),_.textContent=J,this.h()},l(g){t=m(g,"HEADER",{class:!0});var E=H(t);l=m(E,"DIV",{class:!0});var D=H(l);a=m(D,"H2",{class:!0,"data-svelte-h":!0}),w(a)!=="svelte-1t7c000"&&(a.innerHTML=n),i=y(D),o=m(D,"NAV",{id:!0,class:!0});var b=H(o);h=m(b,"A",{href:!0,class:!0,"data-svelte-h":!0}),w(h)!=="svelte-peh6sg"&&(h.textContent=f),e=y(b),s=m(b,"A",{href:!0,class:!0,"data-svelte-h":!0}),w(s)!=="svelte-85qk6k"&&(s.textContent=M),L=y(b),u=m(b,"A",{href:!0,class:!0,"data-svelte-h":!0}),w(u)!=="svelte-j4svvd"&&(u.textContent=T),j=y(b),d=m(b,"A",{href:!0,target:!0,class:!0,"data-svelte-h":!0}),w(d)!=="svelte-h1iwbv"&&(d.textContent=S),x=y(b),_=m(b,"A",{href:!0,target:!0,class:!0,"data-svelte-h":!0}),w(_)!=="svelte-1at6uta"&&(_.textContent=J),b.forEach($),D.forEach($),E.forEach($),this.h()},h(){r(a,"class","brand svelte-jmayj8"),r(h,"href","/"),r(h,"class","svelte-jmayj8"),A(h,"active",v[0]==="/"),r(s,"href","/about"),r(s,"class","svelte-jmayj8"),A(s,"active",v[0]==="/about"),r(u,"href","/projects"),r(u,"class","svelte-jmayj8"),A(u,"active",v[0]==="/projects"),r(d,"href","https://github.com/joshuasand0val"),r(d,"target","_blank"),r(d,"class","svelte-jmayj8"),r(_,"href","https://linkedin.com/in/joshuasand0val/"),r(_,"target","_blank"),r(_,"class","svelte-jmayj8"),r(o,"id","nav"),r(o,"class","svelte-jmayj8"),r(l,"class","wrapper svelte-jmayj8"),r(t,"class","container svelte-jmayj8")},m(g,E){k(g,t,E),c(t,l),c(l,a),c(l,i),c(l,o),c(o,h),c(o,e),c(o,s),c(o,L),c(o,u),c(o,j),c(o,d),c(o,x),c(o,_)},p(g,[E]){E&1&&A(h,"active",g[0]==="/"),E&1&&A(s,"active",g[0]==="/about"),E&1&&A(u,"active",g[0]==="/projects")},i:N,o:N,d(g){g&&$(t)}}}function se(v,t,l){let a,n;return K(v,ee,i=>l(1,n=i)),v.$$.update=()=>{v.$$.dirty&2&&l(0,a=n.url.pathname)},[a,n]}class le extends z{constructor(t){super(),F(this,t,se,ae,q,{})}}function ne(v){let t,l,a,n,i,o=new Date().getFullYear()+"",h,f,e,s,M='Site built with <a target="_blank" href="https://svelte.dev">Svelte</a>.',L,u,T='<i class="bi bi-arrow-up-circle-fill"></i>';return{c(){t=p("footer"),l=p("div"),a=p("div"),n=p("span"),i=V("© "),h=V(o),f=V(" Joshua Elijah Sandoval."),e=C(),s=p("span"),s.innerHTML=M,L=C(),u=p("a"),u.innerHTML=T,this.h()},l(j){t=m(j,"FOOTER",{class:!0});var d=H(t);l=m(d,"DIV",{class:!0});var S=H(l);a=m(S,"DIV",{class:!0});var x=H(a);n=m(x,"SPAN",{class:!0});var _=H(n);i=I(_,"© "),h=I(_,o),f=I(_," Joshua Elijah Sandoval."),_.forEach($),e=y(x),s=m(x,"SPAN",{class:!0,"data-svelte-h":!0}),w(s)!=="svelte-1sf0pv9"&&(s.innerHTML=M),x.forEach($),L=y(S),u=m(S,"A",{href:!0,class:!0,"aria-label":!0,"data-svelte-h":!0}),w(u)!=="svelte-1rnzisb"&&(u.innerHTML=T),S.forEach($),d.forEach($),this.h()},h(){r(n,"class","svelte-1u4c007"),r(s,"class","svelte-1u4c007"),r(a,"class","info svelte-1u4c007"),r(u,"href","#top"),r(u,"class","top svelte-1u4c007"),r(u,"aria-label","To top of page"),A(u,"hidden",!v[0]),r(l,"class","wrapper svelte-1u4c007"),r(t,"class","container svelte-1u4c007")},m(j,d){k(j,t,d),c(t,l),c(l,a),c(a,n),c(n,i),c(n,h),c(n,f),c(a,e),c(a,s),c(l,L),c(l,u)},p(j,[d]){d&1&&A(u,"hidden",!j[0])},i:N,o:N,d(j){j&&$(t)}}}function re(v,t,l){let a=!1;return Q(()=>{const n=document.documentElement,i=()=>{l(0,a=n.scrollHeight>n.offsetHeight&&window.scrollY!==0)};document.addEventListener("scroll",i),window.addEventListener("resize",i)}),[a]}class oe extends z{constructor(t){super(),F(this,t,re,ne,q,{})}}function ie(v){let t,l,a,n,i,o;t=new le({});const h=v[1].default,f=U(h,v,v[0],null);return i=new oe({}),{c(){R(t.$$.fragment),l=C(),a=p("main"),f&&f.c(),n=C(),R(i.$$.fragment),this.h()},l(e){Y(t.$$.fragment,e),l=y(e),a=m(e,"MAIN",{class:!0});var s=H(a);f&&f.l(s),s.forEach($),n=y(e),Y(i.$$.fragment,e),this.h()},h(){r(a,"class","svelte-saovhp")},m(e,s){G(t,e,s),k(e,l,s),k(e,a,s),f&&f.m(a,null),k(e,n,s),G(i,e,s),o=!0},p(e,[s]){f&&f.p&&(!o||s&1)&&W(f,h,e,e[0],o?Z(h,e[0],s,null):X(e[0]),null)},i(e){o||(O(t.$$.fragment,e),O(f,e),O(i.$$.fragment,e),o=!0)},o(e){P(t.$$.fragment,e),P(f,e),P(i.$$.fragment,e),o=!1},d(e){e&&($(l),$(a),$(n)),B(t,e),f&&f.d(e),B(i,e)}}}function ce(v,t,l){let{$$slots:a={},$$scope:n}=t;return v.$$set=i=>{"$$scope"in i&&l(0,n=i.$$scope)},[n,a]}class pe extends z{constructor(t){super(),F(this,t,ce,ie,q,{})}}export{pe as component,de as universal};
