+++ docs/_app/immutable/nodes/1.DmCmqe5i.js
@@ -1,0 +1,1 @@
+import{s as O,c as T}from"../chunks/scheduler.LR2eQXLY.js";import{S as U,i as j,f as k,s as w,e as g,t as q,g as z,j as H,c as $,k as M,l as A,d as h,m as D,a as _,n as F,o as C,b as o,p as L,q as I,r as N,u as P}from"../chunks/index.weWyDj9-.js";import{p as V}from"../chunks/stores.CgJopnDC.js";import{M as B}from"../chunks/Meta.DZWJ86d6.js";function G(c){var x;let a,i,t,r,p=c[0].status+"",d,b,l,u=(((x=c[0].error)==null?void 0:x.message)??"Uhh Oh!")+"",v,E,n,S='Forward to Home <i class="bi bi-chevron-right"></i>',m;return a=new B({props:{title:"Error",description:"Something went wrong... an error occurred."}}),{c(){k(a.$$.fragment),i=w(),t=g("div"),r=g("span"),d=q(p),b=w(),l=g("h1"),v=q(u),E=w(),n=g("a"),n.innerHTML=S,this.h()},l(e){z(a.$$.fragment,e),i=H(e),t=$(e,"DIV",{class:!0});var s=M(t);r=$(s,"SPAN",{class:!0,role:!0});var f=M(r);d=A(f,p),f.forEach(h),b=H(s),l=$(s,"H1",{class:!0});var y=M(l);v=A(y,u),y.forEach(h),E=H(s),n=$(s,"A",{href:!0,"data-svelte-h":!0}),D(n)!=="svelte-zfy13p"&&(n.innerHTML=S),s.forEach(h),this.h()},h(){_(r,"class","status svelte-16mor8m"),_(r,"role","presentation"),_(l,"class","message svelte-16mor8m"),_(n,"href","/"),_(t,"class","container svelte-16mor8m")},m(e,s){F(a,e,s),C(e,i,s),C(e,t,s),o(t,r),o(r,d),o(t,b),o(t,l),o(l,v),o(t,E),o(t,n),m=!0},p(e,[s]){var f;(!m||s&1)&&p!==(p=e[0].status+"")&&L(d,p),(!m||s&1)&&u!==(u=(((f=e[0].error)==null?void 0:f.message)??"Uhh Oh!")+"")&&L(v,u)},i(e){m||(I(a.$$.fragment,e),m=!0)},o(e){N(a.$$.fragment,e),m=!1},d(e){e&&(h(i),h(t)),P(a,e)}}}function J(c,a,i){let t;return T(c,V,r=>i(0,t=r)),[t]}class X extends U{constructor(a){super(),j(this,a,J,G,O,{})}}export{X as component};
