--- docs/_app/immutable/nodes/2.ByGtkBly.js
@@ -1,5 +1,0 @@
-import{s as m,n as p}from"../chunks/scheduler.LR2eQXLY.js";import{S as h,i as v,f as d,s as u,e as f,g as _,j as g,c as j,m as y,a as b,n as k,o,q as $,r as w,d as r,u as x}from"../chunks/index.weWyDj9-.js";import{M as H}from"../chunks/Meta.Bgc6VhtT.js";function S(c){let s,t,a,n=`<img class="profile svelte-rm9oh7" src="pfp.jpg" alt="Joshua Elijah Sandoval"/> <div class="content svelte-rm9oh7"><h1 class="svelte-rm9oh7">Hey, I&#39;m Joshua Elijah Sandoval.</h1> <h2 class="svelte-rm9oh7">A full-stack web developer based in Utah.</h2> <p class="svelte-rm9oh7">Welcome to my website! Here you can explore and learn more
-			<a href="/about" class="svelte-rm9oh7">about myself</a> or <a href="/projects" class="svelte-rm9oh7">my projects</a>, and visit
-			<a href="https://github.com/joshuasand0val" target="_blank" class="svelte-rm9oh7">my Github</a> or
-			<a href="https://linkedin.com/in/joshuasand0val/" target="_blank" class="svelte-rm9oh7">my Linkedin</a>;
-			if perhaps you&#39;re interested.</p> <span class="skills svelte-rm9oh7"><i class="devicon-html5-plain"></i> <i class="devicon-markdown-plain"></i> <i class="devicon-css3-plain"></i> <i class="devicon-sass-plain"></i> <i class="devicon-javascript-plain"></i> <i class="devicon-typescript-plain"></i> <i class="devicon-react-plain"></i> <i class="devicon-svelte-plain"></i> <i class="devicon-eleventy-plain"></i> <i class="devicon-npm-original-wordmark"></i> <i class="devicon-postgresql-plain"></i></span></div>`,i;return s=new H({props:{title:"Home",description:"Joshua Elijah Sandoval is a full-stack web developer."}}),{c(){d(s.$$.fragment),t=u(),a=f("article"),a.innerHTML=n,this.h()},l(e){_(s.$$.fragment,e),t=g(e),a=j(e,"ARTICLE",{class:!0,"data-svelte-h":!0}),y(a)!=="svelte-1bjzo8j"&&(a.innerHTML=n),this.h()},h(){b(a,"class","container svelte-rm9oh7")},m(e,l){k(s,e,l),o(e,t,l),o(e,a,l),i=!0},p,i(e){i||($(s.$$.fragment,e),i=!0)},o(e){w(s.$$.fragment,e),i=!1},d(e){e&&(r(t),r(a)),x(s,e)}}}class q extends h{constructor(s){super(),v(this,s,null,S,m,{})}}export{q as component};
