+++ docs/_app/immutable/nodes/2.C4Aia91M.js
@@ -1,0 +1,5 @@
+import{s as p,n as m}from"../chunks/scheduler.LR2eQXLY.js";import{S as v,i as d,f as u,s as h,e as f,g,j as x,c as _,m as y,a as b,n as k,o,q as j,r as $,d as c,u as w}from"../chunks/index.weWyDj9-.js";import{M as H}from"../chunks/Meta.DZWJ86d6.js";function L(r){let s,t,a,l=`<img class="profile svelte-xeg05n" src="pfp.png" alt="Joshua Elijah Sandoval"/> <div class="content svelte-xeg05n"><h1 class="svelte-xeg05n">Hey, I&#39;m Jes. 💖</h1> <h2 class="svelte-xeg05n">Full-stack web developer. 👩🏻‍💻</h2> <p class="svelte-xeg05n">Welcome to my website! Here you can explore and learn more
+			<a href="/about" class="svelte-xeg05n">about myself</a> or <a href="/projects" class="svelte-xeg05n">my projects</a>, and visit
+			<a href="https://github.com/joshuasand0val" target="_blank" class="svelte-xeg05n">my Github</a> or
+			<a href="https://linkedin.com/in/joshuasand0val/" target="_blank" class="svelte-xeg05n">my Linkedin</a>;
+			if perhaps you&#39;re interested.</p> <span class="skills svelte-xeg05n"><i class="devicon-html5-plain"></i> <i class="devicon-markdown-plain"></i> <i class="devicon-css3-plain"></i> <i class="devicon-sass-plain"></i> <i class="devicon-javascript-plain"></i> <i class="devicon-typescript-plain"></i> <i class="devicon-react-plain"></i> <i class="devicon-svelte-plain"></i> <i class="devicon-eleventy-plain"></i> <i class="devicon-npm-original-wordmark"></i> <i class="devicon-postgresql-plain"></i></span></div>`,n;return s=new H({props:{title:"Home",description:"Joshua Elijah Sandoval is a full-stack web developer."}}),{c(){u(s.$$.fragment),t=h(),a=f("article"),a.innerHTML=l,this.h()},l(e){g(s.$$.fragment,e),t=x(e),a=_(e,"ARTICLE",{class:!0,"data-svelte-h":!0}),y(a)!=="svelte-139lc9g"&&(a.innerHTML=l),this.h()},h(){b(a,"class","container svelte-xeg05n")},m(e,i){k(s,e,i),o(e,t,i),o(e,a,i),n=!0},p:m,i(e){n||(j(s.$$.fragment,e),n=!0)},o(e){$(s.$$.fragment,e),n=!1},d(e){e&&(c(t),c(a)),w(s,e)}}}class C extends v{constructor(s){super(),d(this,s,null,L,p,{})}}export{C as component};
