+++ docs/_app/immutable/nodes/3.CSR92w9f.js
@@ -1,4 +1,4 @@
-import{s as c,n as g}from"../chunks/scheduler.LR2eQXLY.js";import{S as p,i as d,f as u,s as h,e as f,g as v,j as y,c as b,m as w,n as I,o as r,q as _,r as S,d as l,u as k}from"../chunks/index.weWyDj9-.js";import{M as $}from"../chunks/Meta.Bgc6VhtT.js";function x(m){let t,a,n,i=`<h1 class="svelte-v249e3">My name is Joshua Elijah Sandoval.</h1> <h2 class="svelte-v249e3">I am a full-stack web developer.</h2> <div class="container svelte-v249e3"><p>I started my web development journey <strong>back in 2011</strong> after being inspired by
+import{s as c,n as g}from"../chunks/scheduler.LR2eQXLY.js";import{S as p,i as d,f as u,s as h,e as f,g as v,j as y,c as b,m as w,n as I,o as r,q as _,r as k,d as l,u as S}from"../chunks/index.weWyDj9-.js";import{M as $}from"../chunks/Meta.DZWJ86d6.js";function x(m){let t,a,n,i=`<h1 class="svelte-v249e3">My name is Jes.</h1> <h2 class="svelte-v249e3">I am a full-stack web developer.</h2> <div class="container svelte-v249e3"><p>I started my web development journey <strong>back in 2011</strong> after being inspired by
 			the creations of popular YouTube video channels at the time, alongside my father&#39;s notice.
 			He himself being a software architect <b>specializing in the C++ programming language.</b></p> <p>I quickly took an interest to <b>CSS</b>, the technology that builds the internet visually.
 			Since its introduction to me; <strong>it rose to be my absolute favorite.</strong>
@@ -13,4 +13,4 @@ import{s as c,n as g}from"../chunks/scheduler.LR2eQXLY.js";import{S as p,i as d,
 			<a href="https://devmountain.com/" target="_blank">Devmountain</a>.</strong>
 			Beside a cohort of fellow learners and developers, I mastered the art and
 			gained confidence and invaluable insight into the profession.</p> <p>In the now present, I spend my time learning more.
-			<strong>Always ready and able to overcome any new challenges pertaining to web development.</strong></p></div>`,o;return t=new $({props:{title:"About",description:"Learn about myself: Joshua Elijah Sandoval, and my work."}}),{c(){u(t.$$.fragment),a=h(),n=f("article"),n.innerHTML=i},l(e){v(t.$$.fragment,e),a=y(e),n=b(e,"ARTICLE",{"data-svelte-h":!0}),w(n)!=="svelte-1mltpg6"&&(n.innerHTML=i)},m(e,s){I(t,e,s),r(e,a,s),r(e,n,s),o=!0},p:g,i(e){o||(_(t.$$.fragment,e),o=!0)},o(e){S(t.$$.fragment,e),o=!1},d(e){e&&(l(a),l(n)),k(t,e)}}}class L extends p{constructor(t){super(),d(this,t,null,x,c,{})}}export{L as component};
+			<strong>Always ready and able to overcome any new challenges pertaining to web development.</strong></p></div>`,o;return t=new $({props:{title:"About",description:"Learn about myself: Joshua Elijah Sandoval, and my work."}}),{c(){u(t.$$.fragment),a=h(),n=f("article"),n.innerHTML=i},l(e){v(t.$$.fragment,e),a=y(e),n=b(e,"ARTICLE",{"data-svelte-h":!0}),w(n)!=="svelte-e3jxb1"&&(n.innerHTML=i)},m(e,s){I(t,e,s),r(e,a,s),r(e,n,s),o=!0},p:g,i(e){o||(_(t.$$.fragment,e),o=!0)},o(e){k(t.$$.fragment,e),o=!1},d(e){e&&(l(a),l(n)),S(t,e)}}}class L extends p{constructor(t){super(),d(this,t,null,x,c,{})}}export{L as component};
