+++ docs/_app/immutable/nodes/4.Bg-LS3dd.js
@@ -1,0 +1,1 @@
+import{s as D,n as T,d as I}from"../chunks/scheduler.LR2eQXLY.js";import{S as J,i as V,f as x,s as S,e as d,g as B,j as k,c as m,k as _,m as $,d as u,a as n,n as N,o as q,b as c,q as R,r as G,u as U,A as F,t as L,l as z,B as j,C as M}from"../chunks/index.weWyDj9-.js";/* empty css                                    */import{M as K}from"../chunks/Meta.DZWJ86d6.js";function O(a){return(a==null?void 0:a.length)!==void 0?a:Array.from(a)}const Q=[{title:"Hey, Jes!",image:"/projects/joshuasand0val.github.io.png",imageDark:"/projects/joshuasand0val.github.io-dark.png",description:"A personal website showcasing everything about myself, and my work.",url:"/"},{title:"Precolor",description:"A collection of color manipulation tools written in Sass (SCSS).",url:"https://github.com/JoshuaSand0val/Precolor"},{title:"Presized",description:"A collection of resizing tools written in Sass (SCSS).",url:"https://github.com/JoshuaSand0val/Presized"},{title:"Simply Notes!",description:"Simple note taking application with account control.",url:"https://github.com/JesDevBootcamp/simply-notes"}];function W(a,e,s){const t=a.slice();return t[0]=e[s].title,t[1]=e[s].image,t[2]=e[s].imageDark,t[3]=e[s].description,t[4]=e[s].url,t}function X(a){let e,s,t;return{c(){e=j("svg"),s=j("path"),t=j("path"),this.h()},l(r){e=M(r,"svg",{fill:!0,viewBox:!0,class:!0});var i=_(e);s=M(i,"path",{"fill-rule":!0,d:!0}),_(s).forEach(u),t=M(i,"path",{"fill-rule":!0,d:!0}),_(t).forEach(u),i.forEach(u),this.h()},h(){n(s,"fill-rule","evenodd"),n(s,"d","M8.636 3.5a.5.5 0 0 0-.5-.5H1.5A1.5 1.5 0 0 0 0 4.5v10A1.5 1.5 0 0 0 1.5 16h10a1.5 1.5 0 0 0 1.5-1.5V7.864a.5.5 0 0 0-1 0V14.5a.5.5 0 0 1-.5.5h-10a.5.5 0 0 1-.5-.5v-10a.5.5 0 0 1 .5-.5h6.636a.5.5 0 0 0 .5-.5"),n(t,"fill-rule","evenodd"),n(t,"d","M16 .5a.5.5 0 0 0-.5-.5h-5a.5.5 0 0 0 0 1h3.793L6.146 9.146a.5.5 0 1 0 .708.708L15 1.707V5.5a.5.5 0 0 0 1 0z"),n(e,"fill","currentColor"),n(e,"viewBox","-16 -16 48 48"),n(e,"class","svelte-ai4q0w")},m(r,i){q(r,e,i),c(e,s),c(e,t)},d(r){r&&u(e)}}}function Y(a){let e,s;return{c(){e=d("img"),this.h()},l(t){e=m(t,"IMG",{src:!0,loading:!0,alt:!0,class:!0}),this.h()},h(){I(e.src,s=a[1])||n(e,"src",s),n(e,"loading","lazy"),n(e,"alt",""),n(e,"class","svelte-ai4q0w")},m(t,r){q(t,e,r)},d(t){t&&u(e)}}}function Z(a){let e,s=`<source srcset="${a[2]}" media="(prefers-color-scheme: dark)"/> <img src="${a[1]}" loading="lazy" alt="" class="svelte-ai4q0w"/>`;return{c(){e=d("picture"),e.innerHTML=s},l(t){e=m(t,"PICTURE",{"data-svelte-h":!0}),$(e)!=="svelte-1yiwr41"&&(e.innerHTML=s)},m(t,r){q(t,e,r)},d(t){t&&u(e)}}}function ee(a){let e,s,t,r,i,w,h,b,y,o,l=`<a href="${a[4]}" target="_blank">Learn More <i class="bi bi-chevron-right"></i></a>`,f;function v(g,p){return[g[1],g[2]].every(se)?Z:typeof g[1]=="string"?Y:X}let E=v(a)(a);return{c(){e=d("article"),s=d("h2"),t=L(a[0]),r=S(),i=d("a"),E.c(),w=S(),h=d("p"),b=L(a[3]),y=S(),o=d("span"),o.innerHTML=l,f=S(),this.h()},l(g){e=m(g,"ARTICLE",{class:!0});var p=_(e);s=m(p,"H2",{class:!0});var A=_(s);t=z(A,a[0]),A.forEach(u),r=k(p),i=m(p,"A",{class:!0,href:!0,target:!0});var P=_(i);E.l(P),P.forEach(u),w=k(p),h=m(p,"P",{class:!0});var H=_(h);b=z(H,a[3]),H.forEach(u),y=k(p),o=m(p,"SPAN",{class:!0,"data-svelte-h":!0}),$(o)!=="svelte-zs1twe"&&(o.innerHTML=l),f=k(p),p.forEach(u),this.h()},h(){n(s,"class","title svelte-ai4q0w"),n(i,"class","portal svelte-ai4q0w"),n(i,"href",a[4]),n(i,"target","_blank"),n(h,"class","description svelte-ai4q0w"),n(o,"class","learn-more svelte-ai4q0w"),n(e,"class","project svelte-ai4q0w")},m(g,p){q(g,e,p),c(e,s),c(s,t),c(e,r),c(e,i),E.m(i,null),c(e,w),c(e,h),c(h,b),c(e,y),c(e,o),c(e,f)},p:T,d(g){g&&u(e),E.d()}}}function te(a){let e,s,t,r,i="My Projects",w,h,b;e=new K({props:{title:"Projects",description:"The various projects undertaken by Joshua Elijah Sandoval."}});let y=O(Q),o=[];for(let l=0;l<y.length;l+=1)o[l]=ee(W(a,y,l));return{c(){x(e.$$.fragment),s=S(),t=d("div"),r=d("h1"),r.textContent=i,w=S(),h=d("div");for(let l=0;l<o.length;l+=1)o[l].c();this.h()},l(l){B(e.$$.fragment,l),s=k(l),t=m(l,"DIV",{class:!0});var f=_(t);r=m(f,"H1",{class:!0,"data-svelte-h":!0}),$(r)!=="svelte-rvn3aq"&&(r.textContent=i),w=k(f),h=m(f,"DIV",{class:!0});var v=_(h);for(let C=0;C<o.length;C+=1)o[C].l(v);v.forEach(u),f.forEach(u),this.h()},h(){n(r,"class","svelte-ai4q0w"),n(h,"class","grid svelte-ai4q0w"),n(t,"class","container")},m(l,f){N(e,l,f),q(l,s,f),q(l,t,f),c(t,r),c(t,w),c(t,h);for(let v=0;v<o.length;v+=1)o[v]&&o[v].m(h,null);b=!0},p:T,i(l){b||(R(e.$$.fragment,l),b=!0)},o(l){G(e.$$.fragment,l),b=!1},d(l){l&&(u(s),u(t)),U(e,l),F(o,l)}}}const se=a=>typeof a=="string";class ie extends J{constructor(e){super(),V(this,e,null,te,D,{})}}export{ie as component};
