--- docs/_app/immutable/nodes/4.CgI26PQP.js
@@ -1,1 +1,0 @@
-import{s as V,n as D,d as P}from"../chunks/scheduler.LR2eQXLY.js";import{S as q,i as B,f as T,s as S,e as p,g as x,j as k,c as v,k as b,m as J,d as h,a,n as G,o as y,b as m,q as N,r as R,u as F,A as K,t as I,l as L,B as A,C}from"../chunks/index.weWyDj9-.js";/* empty css                                    */import{M as O}from"../chunks/Meta.Bgc6VhtT.js";function Q(r){return(r==null?void 0:r.length)!==void 0?r:Array.from(r)}const U=[{title:"Hey, Jes!",image:"/projects/joshuasand0val.github.io.png",imageDark:"/projects/joshuasand0val.github.io-dark.png",description:"A personal website showcasing everything about myself, and my work.",url:"/"},{title:"Precolor",description:"A collection of color manipulation tools written in Sass (SCSS).",url:"https://github.com/JoshuaSand0val/Precolor"},{title:"Presized",description:"A collection of resizing tools written in Sass (SCSS).",url:"https://github.com/JoshuaSand0val/Presized"},{title:"Simply Notes!",description:"Simple note taking application with account control.",url:"https://github.com/JesDevBootcamp/simply-notes"}];function W(r,e,s){const t=r.slice();return t[0]=e[s].title,t[1]=e[s].image,t[2]=e[s].imageDark,t[3]=e[s].description,t[4]=e[s].url,t}function X(r){let e,s,t;return{c(){e=A("svg"),s=A("path"),t=A("path"),this.h()},l(l){e=C(l,"svg",{fill:!0,viewBox:!0,class:!0});var i=b(e);s=C(i,"path",{"fill-rule":!0,d:!0}),b(s).forEach(h),t=C(i,"path",{"fill-rule":!0,d:!0}),b(t).forEach(h),i.forEach(h),this.h()},h(){a(s,"fill-rule","evenodd"),a(s,"d","M8.636 3.5a.5.5 0 0 0-.5-.5H1.5A1.5 1.5 0 0 0 0 4.5v10A1.5 1.5 0 0 0 1.5 16h10a1.5 1.5 0 0 0 1.5-1.5V7.864a.5.5 0 0 0-1 0V14.5a.5.5 0 0 1-.5.5h-10a.5.5 0 0 1-.5-.5v-10a.5.5 0 0 1 .5-.5h6.636a.5.5 0 0 0 .5-.5"),a(t,"fill-rule","evenodd"),a(t,"d","M16 .5a.5.5 0 0 0-.5-.5h-5a.5.5 0 0 0 0 1h3.793L6.146 9.146a.5.5 0 1 0 .708.708L15 1.707V5.5a.5.5 0 0 0 1 0z"),a(e,"fill","currentColor"),a(e,"viewBox","-16 -16 48 48"),a(e,"class","svelte-1ymw1bw")},m(l,i){y(l,e,i),m(e,s),m(e,t)},d(l){l&&h(e)}}}function Y(r){let e,s;return{c(){e=p("img"),this.h()},l(t){e=v(t,"IMG",{src:!0,loading:!0,alt:!0,class:!0}),this.h()},h(){P(e.src,s=r[1])||a(e,"src",s),a(e,"loading","lazy"),a(e,"alt",""),a(e,"class","svelte-1ymw1bw")},m(t,l){y(t,e,l)},d(t){t&&h(e)}}}function Z(r){let e,s,t,l,i;return{c(){e=p("img"),t=S(),l=p("img"),this.h()},l(o){e=v(o,"IMG",{src:!0,loading:!0,class:!0,alt:!0}),t=k(o),l=v(o,"IMG",{src:!0,loading:!0,class:!0,alt:!0}),this.h()},h(){P(e.src,s=r[1])||a(e,"src",s),a(e,"loading","lazy"),a(e,"class","system light svelte-1ymw1bw"),a(e,"alt",""),P(l.src,i=r[2])||a(l,"src",i),a(l,"loading","lazy"),a(l,"class","system dark svelte-1ymw1bw"),a(l,"alt","")},m(o,c){y(o,e,c),y(o,t,c),y(o,l,c)},d(o){o&&(h(e),h(t),h(l))}}}function ee(r){let e,s,t,l,i,o,c,w,j,u,n=`<a href="${r[4]}" target="_blank">Learn More <i class="bi bi-chevron-right"></i></a>`,f;function g(_,d){return[_[1],_[2]].every(se)?Z:typeof _[1]=="string"?Y:X}let M=g(r)(r);return{c(){e=p("article"),s=p("h2"),t=I(r[0]),l=S(),i=p("a"),M.c(),o=S(),c=p("p"),w=I(r[3]),j=S(),u=p("span"),u.innerHTML=n,f=S(),this.h()},l(_){e=v(_,"ARTICLE",{class:!0});var d=b(e);s=v(d,"H2",{class:!0});var $=b(s);t=L($,r[0]),$.forEach(h),l=k(d),i=v(d,"A",{class:!0,href:!0,target:!0});var z=b(i);M.l(z),z.forEach(h),o=k(d),c=v(d,"P",{class:!0});var H=b(c);w=L(H,r[3]),H.forEach(h),j=k(d),u=v(d,"SPAN",{class:!0,"data-svelte-h":!0}),J(u)!=="svelte-zs1twe"&&(u.innerHTML=n),f=k(d),d.forEach(h),this.h()},h(){a(s,"class","title svelte-1ymw1bw"),a(i,"class","portal svelte-1ymw1bw"),a(i,"href",r[4]),a(i,"target","_blank"),a(c,"class","description svelte-1ymw1bw"),a(u,"class","learn-more svelte-1ymw1bw"),a(e,"class","project svelte-1ymw1bw")},m(_,d){y(_,e,d),m(e,s),m(s,t),m(e,l),m(e,i),M.m(i,null),m(e,o),m(e,c),m(c,w),m(e,j),m(e,u),m(e,f)},p:D,d(_){_&&h(e),M.d()}}}function te(r){let e,s,t,l,i="My Projects",o,c,w;e=new O({props:{title:"Projects",description:"The various projects undertaken by Joshua Elijah Sandoval."}});let j=Q(U),u=[];for(let n=0;n<j.length;n+=1)u[n]=ee(W(r,j,n));return{c(){T(e.$$.fragment),s=S(),t=p("div"),l=p("h1"),l.textContent=i,o=S(),c=p("div");for(let n=0;n<u.length;n+=1)u[n].c();this.h()},l(n){x(e.$$.fragment,n),s=k(n),t=v(n,"DIV",{class:!0});var f=b(t);l=v(f,"H1",{class:!0,"data-svelte-h":!0}),J(l)!=="svelte-rvn3aq"&&(l.textContent=i),o=k(f),c=v(f,"DIV",{class:!0});var g=b(c);for(let E=0;E<u.length;E+=1)u[E].l(g);g.forEach(h),f.forEach(h),this.h()},h(){a(l,"class","svelte-1ymw1bw"),a(c,"class","grid svelte-1ymw1bw"),a(t,"class","container")},m(n,f){G(e,n,f),y(n,s,f),y(n,t,f),m(t,l),m(t,o),m(t,c);for(let g=0;g<u.length;g+=1)u[g]&&u[g].m(c,null);w=!0},p:D,i(n){w||(N(e.$$.fragment,n),w=!0)},o(n){R(e.$$.fragment,n),w=!1},d(n){n&&(h(s),h(t)),F(e,n),K(u,n)}}}const se=r=>typeof r=="string";class ie extends q{constructor(e){super(),B(this,e,null,te,V,{})}}export{ie as component};
